Implemented State Machine for Combat (mostly)
All state machine scripts are contained in the StateMachine folder, they mainly consist of replacements for existing scripts. They were made separate for testing purposes, in case this breaks the game beyond repair. None of the objects in the actual scene have any of the state machine scripts attached, and none of these are tested. Use at your own peril.